requirements.txt
aio-pika==9.4.3
aiormq==6.8.1
asyncio==3.4.3
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
multidict==6.1.0
pamqp==3.3.0
python-dotenv==1.0.1
requests==2.32.3
tapo==0.5.1
urllib3==2.2.3
yarl==1.13.1
